requirements.txt
amqp==5.0.6
asgiref==3.3.4
billiard==3.6.4.0
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
colorama==0.4.4
configparser==5.0.2
contextvars==2.4
crayons==0.4.0
Django==3.2
django-filter==2.4.0
djangorestframework==3.12.4
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2020.12.22
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
immutables==0.15
importlib-metadata==4.0.1
kombu==5.0.2
lxml==4.6.3
Markdown==3.3.4
prompt-toolkit==3.0.18
pytz==2021.1
requests==2.25.1
rfc3986==1.4.0
six==1.15.0
sniffio==1.2.0
soupsieve==2.2.1
sqlparse==0.4.1
typing-extensions==3.7.4.3
urllib3==1.26.4
vine==5.0.0
wcwidth==0.2.5
webdriver-manager==3.4.0
zipp==3.4.1
psycopg2==2.8.6